+++ rust/connlib/tunnel/src/device_channel.rs
@@ -28,11 +28,25 @@ pub struct Device {
 }
 
 impl Device {
+    #[cfg(target_family = "unix")]
     pub(crate) async fn new(
         config: &Interface,
         callbacks: &impl Callbacks<Error = Error>,
     ) -> Result<Device, ConnlibError> {
-        create_iface(config, callbacks).await
+        let (iface, stream) = IfaceDevice::new(config, callbacks).await?;
+        iface.up().await?;
+        let io = DeviceIo(stream);
+        let mtu = AtomicUsize::new(ioctl::interface_mtu_by_name(iface.name())?);
+
+        Ok(Device { io, mtu, iface })
+    }
+
+    #[cfg(target_family = "windows")]
+    pub(crate) async fn new(
+        config: &Interface,
+        callbacks: &impl Callbacks<Error = Error>,
+    ) -> Result<Device, ConnlibError> {
+        todo!()
     }
 
     pub(crate) fn poll_read<'b>(
